# src/main/resources/swagger/swagger-workstation.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: API Workstation
  description: 'Hackathon'
  contact:
    name: Team 3 - GFT Brasil
    email: andu@gft.com
host: gft.com.br
basePath: /gft/workstation/v1/office/
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /office:
    post:
      tags:
        - Office
      summary: Create a office
      description: 'post a office'
      operationId: createOffice
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: Office
          in: body
          description: office object
          required: true
          schema:
            $ref: '#/definitions/Office'
      responses:
        '201':
          description: Create new resource
          schema:
            $ref: '#/definitions/Office'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'

    get:
      tags:
        - Office
      summary: Gets all office
      description: 'get a office'
      operationId: getOffice
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OfficesResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /office/{id}:
    get:
      tags:
        - Office
      summary: Get office
      description: 'get a office'
      operationId: getOfficeById
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of workstation
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OfficeResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'

    delete:
      tags:
        - Office
      summary: Delete office
      description: 'delete a office'
      operationId: deleteOffice
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of workstation
          required: true
          type: string
          x-example: '000001'
      responses:
        '204':
          description: No Content
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /booking:
    post:
      tags:
        - Bookings
      summary: Create a or a lot bookings
      description: 'put a user'
      operationId: createBookings
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: Booking
          in: body
          description: Booking object
          required: true
          schema:
            $ref: '#/definitions/Bookings'
      responses:
        '201':
          description: Create new resource
          schema:
            $ref: '#/definitions/BookingResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{id}:
    get:
      tags:
        - Bookings
      summary: Gets a booking
      description: 'get a booking'
      operationId: getBookingsById
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of booking
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/BookingResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      tags:
        - Bookings
      summary: Update a booking
      description: 'put a user'
      operationId: updateBooking
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: booking
          required: true
          type: string
          x-example: 'XPTO'
        - name: Booking
          in: body
          description: Booking object
          required: true
          schema:
            $ref: '#/definitions/Booking'
      responses:
        '204':
          description: No Content

        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{idUser}/user:
    get:
      tags:
        - Bookings
      summary: Gets Bookings of the User
      description: 'Get bookings of the User'
      operationId: getBookingsByIdUser
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: idUser
          in: path
          description: Code of booking
          required: true
          type: string
          x-example: '000001'
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/BookingsResponse'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Document not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{id}/checkin:
    patch:
      tags:
        - Bookings
      summary: Update a user
      description: 'put a user'
      operationId: createCheckin
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: booking
          required: true
          type: string
          x-example: 'XPTO'
        - name: Booking
          in: body
          description: data of the Booking
          required: true
          schema:
            $ref: '#/definitions/CheckinCheckout'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
  /booking/{id}/checkout:
    patch:
      tags:
        - Bookings
      summary: Update a user
      description: 'put a user'
      operationId: createCheckout
      produces:
        - application/json
      security:
        - OauthSecurity:
            - bearer
      parameters:
        - name: id
          in: path
          description: Code of the Booking
          required: true
          type: string
          x-example: 'XPTO'
        - name: Booking
          in: body
          description: data of the Booking
          required: true
          schema:
            $ref: '#/definitions/CheckinCheckout'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorModel'
...
securityDefinitions:
  OauthSecurity:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    tokenUrl: 'https://sandbox-api.exemple.com.br/security/iam/v1/user-identities/login'
    scopes:
      bearer: Client authorization token - Bearer JWT
security:
  - OauthSecurity:
      - bearer
definitions:
  Office:
    description: Office
    type: object
    properties:
      id:
        type: string
        example: 0001
        description: Unique identifier of the Office
      name:
        type: string
        example: Escritorio Sorocaba
        description: Name
      address:
        type: string
        example: Sorocaba
        description: Endereco
      status:
        type: string
        example: ACTIVE
        enum: [ACTIVE, INACTIVE]
  OfficeResponse:
    description: Office
    type: object
    properties:
      id:
        type: string
        example: 0001
        description: Unique identifier of the Office
      name:
        type: string
        example: Escritorio Sorocaba
        description: Name
      address:
        type: string
        example: Sorocaba
        description: Endereco
      createdDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
      status:
        type: string
        example: ACTIVE
        enum: [ACTIVE, INACTIVE]
  OfficesResponse:
    type: object
    required:
      - office
    properties:
      offices:
        type: array
        items:
          $ref: '#/definitions/OfficeResponse'
...
  Booking:
    description: Booking
    type: object
    required:
      - idUser
      - idWorkstation
    properties:
      idUser:
        type: string
        example: 001
        description: Id of the User
      idWorkstation:
        type: string
        example: 001
        description: Id of the Workstation
      idOffice:
        type: string
        example: 001
        description: Id of the Office
      bookingDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Start Date of the validity of the Booking.
      checkinDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkIn.
      checkoutDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkOut.
      sizeHours:
        type: string
        format: int
        example: 4
        description: Hours.
      startedTime:
        type: string
        example: '10:00'
        description: Hours.
      finishedTime:
        type: string
        example: '12:00'
        description: Hours.
      qrCode:
        type: string
        example: '12:00'
        description: Hours.
      status:
        type: string
        enum: [ACTIVE, INACTIVE, CANCELED, FINISHED, IN_USE, RESERVED, BLOCKED]
  Bookings:
    type: object
    properties:
      bookings:
        type: array
        items:
          $ref: '#/definitions/Booking'
  BookingResponse:
    description: Booking
    type: object
    properties:
      id:
        type: string
        example: 123
        description: Id of the Booking
      idUser:
        type: string
        example: string
        description: Id of the User
      idWorkstation:
        type: string
        example: string
        description: Id of the Workstation
      idOffice:
        type: string
        example: andu
        description: Id of the Office
      bookingDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Start Date of the validity of the Booking.
      checkinDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkIn.
      checkoutDate:
        type: string
        format: date-time
        example: yyyy-mm-dd
        description: Data do checkOut.
      startedDatetime:
        type: string
        example: '09:00'
        description: Data do checkOut.
      finishedDatetime:
        type: string
        example: '12:00'
        description: Data do checkOut.
      sizeHours:
        type: string
        format: int64
        example: 4
        description: Start Date of the validity of the Booking.
      codeBar:
        type: string
        example: 123456
      status:
        type: string
        enum: [ACTIVE, INACTIVE, CANCELED, FINISHED, IN_USE, RESERVED, BLOCKED, OTHERS]
  BookingsResponse:
    type: object
    properties:
      bookings:
        type: array
        items:
          $ref: '#/definitions/BookingResponse'
  CheckinCheckout:
    description: Checkin
    type: object
    required:
      - id
    properties:
      qrCode:
        type: string
        example: 001
        description: QR Code
...
  ErrorModel:
    type: object
    required:
      - codeBuilderComplete
      - message
    properties:
      code:
        type: string
      message:
        type: string